+++ models/projeto_subsidio_sppo_encontro_contas/CHANGELOG.md
@@ -8,7 +8,7 @@
 
 ### Alterado
 
-- Altera e corrige tratamento de serviços do RDO no modelo `balanco_servico_dia.sql` (https://github.com/prefeitura-rio/queries-rj-smtr/pull/332)
+- Altera e corrige tratamento de serviços do RDO nos modelos `balanco_servico_dia.sql` e `aux_balanco_rdo_servico_dia.sql` (https://github.com/prefeitura-rio/queries-rj-smtr/pull/332)
 
 ## [1.0.2] - 2024-05-21